requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
groq==0.4.1
transformers==4.36.0
sentence-transformers==2.2.2
chromadb==0.4.18
pypdf==3.17.0
python-docx==1.1.0
Werkzeug==2.3.7